+++ lib/src/native.dart
@@ -5,6 +5,7 @@ import 'package:flutter/foundation.dart';
 import 'package:layrz_logging/src/models.dart';
 import 'package:layrz_theme/layrz_theme.dart';
 import 'package:path_provider/path_provider.dart';
+import 'package:permission_handler/permission_handler.dart';
 
 Future<void> initLogFile() async {
   // Ensure log file is initialized
@@ -22,7 +23,7 @@ Future<void> initLogFile() async {
 
 Future<void> saveIntoFile(Log log) async {
   // Save log into a file
-  final Directory directory = await getApplicationSupportDirectory();
+  final Directory directory = await getApplicationDocumentsDirectory();
   File file = File('${directory.path}/latest.log');
 
   file.writeAsStringSync(
@@ -31,12 +32,39 @@ Future<void> saveIntoFile(Log log) async {
   );
 }
 
-Future<void> openLogFile() async {
+Future<String?> openLogFile() async {
   try {
-    final Directory directory = await getApplicationSupportDirectory();
+    final lines = await fetchLogsFromFile();
 
-    List<String> lines = [];
+    final res = await Permission.storage.status;
+    if (res == PermissionStatus.denied) {
+      final status = await Permission.storage.request();
+
+      if (status != PermissionStatus.granted) {
+        return null;
+      }
+    } else if (res == PermissionStatus.permanentlyDenied) {
+      return null;
+    }
+
+    final output = await saveFile(
+      filename: 'export.log',
+      bytes: utf8.encode(lines.join('\n')),
+    );
+    return output?.path;
+  } catch (e) {
+    debugPrint('Error opening log file: $e');
+    return null;
+  }
+}
+
+Future<List<String>> fetchLogsFromFile() async {
+  List<String> lines = [];
+
+  try {
+    final Directory directory = await getApplicationDocumentsDirectory();
     File file = File('${directory.path}/latest.log');
+
     if (file.existsSync()) {
       lines.add("------- LATEST LOG ------");
       lines.addAll(file.readAsLinesSync());
@@ -47,12 +75,9 @@ Future<void> openLogFile() async {
       lines.add("------- PREVIOUS LOG ------");
       lines.addAll(file.readAsLinesSync());
     }
-
-    await saveFile(
-      filename: 'latest.log',
-      bytes: utf8.encode(lines.join('\n')),
-    );
   } catch (e) {
-    debugPrint('Error opening log file: $e');
+    debugPrint('Error fetching logs from file: $e');
   }
+
+  return lines;
 }