+++ lib/src/web.dart
@@ -1,7 +1,17 @@
 import 'package:layrz_logging/src/models.dart';
 
-Future<void> initLogFile() async {}
+Future<void> initLogFile() async {
+  throw UnsupportedError("This method is not supported on the web");
+}
 
-Future<void> saveIntoFile(Log log) async {}
+Future<void> saveIntoFile(Log log) async {
+  throw UnsupportedError("This method is not supported on the web");
+}
 
-Future<void> openLogFile() async {}
+Future<String?> openLogFile() async {
+  throw UnsupportedError("This method is not supported on the web");
+}
+
+Future<List<String>> fetchLogsFromFile() async {
+  throw UnsupportedError("This method is not supported on the web");
+}