+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 1.0.8
+
+- Added new method to readLogs from the files created, and return a `List<String>`.
+- Changed method `openLogfile` to use the same method that `readLogs`.
+- Added `PreviewLogDialog` widget to show the logs in a dialog.
+- Changed `openLogfile` to validate the permissions before trying to save a file.
+
+## 1.0.7
+
+- Added `openLogfile` method to open the log file in the default text editor.
+
 ## 1.0.6
 
 - Removed `open_file` dependency to support WASM compilation.